Add README and reference it in setup.py
Adding a readme that will serve on both github and PYPI, so format is an
RST file, and formatting within attempts to use syntax that overlaps
both markdown and restructured text. Also made changes to setup.py to
reference this as the long description so it shows up on the PYPI page
for it.